+++ Mailer/app/Models/PageMeta/Content/EuobserverContent.php
@@ -5,22 +5,16 @@ namespace Remp\Mailer\Models\PageMeta\Content;
 
 use Nette\Utils\Strings;
 use Remp\MailerModule\Models\PageMeta\Content\JsonLDContent;
-use Remp\MailerModule\Models\PageMeta\Meta;
 
 class EuobserverContent extends JsonLDContent
 {
-    protected function postProcessMeta(Meta $meta): Meta
+    protected function processAuthors(array $authors): array
     {
-        $authors = [];
-        foreach ($meta->getAuthors() as $author) {
-            $authors[] = Strings::upper($author);
+        $upperCasedAuthors = [];
+        foreach ($authors as $author) {
+            $upperCasedAuthors[] = Strings::upper($author);
         }
 
-        return new Meta(
-            $meta->getTitle(),
-            $meta->getDescription(),
-            $meta->getImage(),
-            $authors,
-        );
+        return $upperCasedAuthors;
     }
 }